Updated autotests for PublicIP
