+++ test/commands/arm/network/arm.network.public-ip-tests.js
@@ -23,6 +23,7 @@
 var should = require('should');
 var util = require('util');
 var _ = require('underscore');
+var fs = require('fs');
 
 var CLITest = require('../../../framework/arm-cli-test');
 var utils = require('../../../../lib/util/utils');
@@ -30,71 +31,55 @@ var tagUtils = require('../../../../lib/commands/arm/tag/tagUtils');
 var testUtils = require('../../../util/util');
 
 var networkTestUtil = new (require('../../../util/networkTestUtil'))();
+var vmTestUtil = new (require('../../../util/vmTestUtil'))();
 
 var generatorUtils = require('../../../../lib/util/generatorUtils');
 var profile = require('../../../../lib/util/profile');
 var $ = utils.getLocaleString;
 
 var testPrefix = 'arm-network-public-ip-tests',
-  groupName = 'xplat-test-public-ip',
-  location;
-var index = 0;
+  groupName = 'xplat-test-public-ip-custom',
+  location,
+  vmssParams;
 
 var publicIPAddresses = {
   publicIPAllocationMethod: 'Static',
-  publicIPAllocationMethodNew: 'Dynamic',
   publicIPAddressVersion: 'IPv4',
   domainNameLabel: 'labelcreate',
-  domainNameLabelNew: 'labelset',
-  idleTimeoutInMinutes: '15',
-  idleTimeoutInMinutesNew: '14',
+  idleTimeoutInMinutes: 15,
   location: 'southeastasia',
+  virtualmachineIndex: '0',
+  networkInterfaceName: 'test',
+  ipConfigurationName: 'test',
+  publicIpAddressName: 'xplattestip',
   name: 'publicIPAddressName'
 };
-var publicIPAddressesDefault = {
-  location: 'southeastasia',
-  publicIPAllocationMethod: 'Dynamic',
-  publicIPAddressVersion: 'IPv4',
-  idleTimeoutInMinutes: '4',
-  name: 'publicIPAddressesDefaultName',
-  group: groupName
+
+publicIPAddresses.virtualNetworkName = 'virtualNetworkName';
+publicIPAddresses.subnetName = 'subnetName';
+publicIPAddresses.networkSecurityGroupName = 'networkSecurityGroupName';
+
+var subnet = {
+  name: 'subnetName',
+  addressPrefix: '10.0.0.0/16'
 };
-var ipAllocationMethodOutOfRange = {
-  publicIPAllocationMethod: 'Any',
-  location: 'southeastasia',
-  name: 'IPAllocationMethodOutOfRangeName',
-  group: groupName
+var virtualNetwork = {
+  name: 'virtualNetworkName',
+  location: 'southeastasia'
 };
-var ipVersionOutOfRange = {
-  publicIPAddressVersion: 'IPv8',
-  location: 'southeastasia',
-  name: 'IPVersionOutOfRangeName',
-  group: groupName
+var networkSecurityGroup = {
+  name: 'networkSecurityGroupName',
+  location: 'southeastasia'
 };
-var invalidSymbolsInLable = {
-  domainNameLabel: 'l-a_b-1-e',
-  location: 'southeastasia',
-  name: 'invalidSymbolsInLableName',
-  group: groupName
-};
-var idleTimeoutUnderRange = {
-  idleTimeoutInMinutes: '3',
-  location: 'southeastasia',
-  name: 'idleTimeoutUnderRangeName',
-  group: groupName
-};
-var idleTimeoutOverRange = {
-  idleTimeoutInMinutes: '31',
-  location: 'southeastasia',
-  name: 'idleTimeoutOverRangeName',
-  group: groupName
-};
-var deleteOfDomainNameLabel = {
-  domainNameLabel: 'xplattestlbl',
-  domainNameLabelNew: '',
-  location: 'southeastasia',
-  name: 'deleteOfDomainNameLabelName',
-  group: groupName
+var vmTestUtilCreateVMSSWithParamFile = {
+  vmssName: 'publicIPAddressName',
+  paramFile: 'test/data/vmssParamTestPublicIp.json',
+  storageAccount: 'xplatteststorage1',
+  storageCont: 'xplatteststoragecnt1',
+  publicIpName: 'xplattestip',
+  domainNameLabel: 'vmssiplabel',
+  idleTimeout: '30',
+  nicDnsServer: '10.11.12.13'
 };
 
 var requiredEnvironment = [{
@@ -106,28 +91,56 @@ describe('arm', function () {
   describe('network', function () {
     var suite, retry = 5;
     var hour = 60 * 60000;
+    var testTimeout = hour;
 
     before(function (done) {
-      this.timeout(hour);
+      this.timeout(testTimeout);
       suite = new CLITest(this, testPrefix, requiredEnvironment);
       suite.setupSuite(function () {
         location = publicIPAddresses.location || process.env.AZURE_VM_TEST_LOCATION;
         groupName = suite.isMocked ? groupName : suite.generateId(groupName, null);
+        
         publicIPAddresses.location = location;
         publicIPAddresses.group = groupName;
         publicIPAddresses.name = suite.isMocked ? publicIPAddresses.name : suite.generateId(publicIPAddresses.name, null);
+
+        vmssParams = JSON.parse(fs.readFileSync(vmTestUtilCreateVMSSWithParamFile.paramFile));
+
         if (!suite.isPlayback()) {
           networkTestUtil.createGroup(groupName, location, suite, function () {
-            done();
+            var cmd = 'network vnet create -g {1} -n {name} --location {location} --json'.formatArgs(virtualNetwork, groupName);
+            testUtils.executeCommand(suite, retry, cmd, function (result) {
+              result.exitStatus.should.equal(0);
+              var output = JSON.parse(result.text);
+              publicIPAddresses.virtualNetworkId = output.id;
+
+              var cmd = 'network vnet subnet create -g {1} -n {name} --address-prefix {addressPrefix} --vnet-name virtualNetworkName --json'.formatArgs(subnet, groupName);
+              testUtils.executeCommand(suite, retry, cmd, function (result) {
+                result.exitStatus.should.equal(0);
+                var output = JSON.parse(result.text);
+                publicIPAddresses.subnetId = output.id;
+
+                var cmd = 'network nsg create -g {1} -n {name} --location {location} --json'.formatArgs(networkSecurityGroup, groupName);
+                testUtils.executeCommand(suite, retry, cmd, function (result) {
+                  result.exitStatus.should.equal(0);
+                  var output = JSON.parse(result.text);
+                  publicIPAddresses.networkSecurityGroupId = output.id;
+                  done();
+                });
+              });
+            });
           });
         } else {
           var subscriptionId = profile.current.getSubscription().id;
+          publicIPAddresses.virtualNetworkId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks', virtualNetwork.name);
+          publicIPAddresses.subnetId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'virtualNetworks/' + virtualNetwork.name + '/subnets', subnet.name);
+          publicIPAddresses.networkSecurityGroupId = generatorUtils.generateResourceIdCommon(subscriptionId, groupName, 'networkSecurityGroups', networkSecurityGroup.name);
           done();
         }
       });
     });
     after(function (done) {
-      this.timeout(hour);
+      this.timeout(testTimeout);
       networkTestUtil.deleteGroup(groupName, suite, function () {
         suite.teardownSuite(done);
       });
@@ -139,57 +152,57 @@ describe('arm', function () {
       suite.teardownTest(done);
     });
 
-    describe('public ip addresses', function () {
-      this.timeout(hour);
+    describe('public ip addresses custom', function () {
+      this.timeout(testTimeout);
+
       it('create should create public ip addresses', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --allocation-method {publicIPAllocationMethod} --ip-version {publicIPAddressVersion} --domain-name-label {domainNameLabel} --idle-timeout {idleTimeoutInMinutes} --location {location} --json'.formatArgs(publicIPAddresses);
+        var cmd = ('network public-ip create -g {group} -n {name} --allocation-method {publicIPAllocationMethod} ' +
+          '--ip-version {publicIPAddressVersion} --domain-name-label {domainNameLabel} ' +
+          '--idle-timeout {idleTimeoutInMinutes} --location {location} --json').formatArgs(publicIPAddresses);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
+
           output.name.should.equal(publicIPAddresses.name);
           output.publicIPAllocationMethod.toLowerCase().should.equal(publicIPAddresses.publicIPAllocationMethod.toLowerCase());
           output.publicIPAddressVersion.toLowerCase().should.equal(publicIPAddresses.publicIPAddressVersion.toLowerCase());
           output.dnsSettings.domainNameLabel.toLowerCase().should.equal(publicIPAddresses.domainNameLabel.toLowerCase());
-          output.idleTimeoutInMinutes.should.equal(parseInt(publicIPAddresses.idleTimeoutInMinutes, 10));
+          output.idleTimeoutInMinutes.should.equal(publicIPAddresses.idleTimeoutInMinutes);
+
           done();
         });
       });
+
       it('show should display public ip addresses details', function (done) {
         var cmd = 'network public-ip show -g {group} -n {name} --json'.formatArgs(publicIPAddresses);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
+
           output.name.should.equal(publicIPAddresses.name);
           output.publicIPAllocationMethod.toLowerCase().should.equal(publicIPAddresses.publicIPAllocationMethod.toLowerCase());
           output.publicIPAddressVersion.toLowerCase().should.equal(publicIPAddresses.publicIPAddressVersion.toLowerCase());
           output.dnsSettings.domainNameLabel.toLowerCase().should.equal(publicIPAddresses.domainNameLabel.toLowerCase());
-          output.idleTimeoutInMinutes.should.equal(parseInt(publicIPAddresses.idleTimeoutInMinutes, 10));
+          output.idleTimeoutInMinutes.should.equal(publicIPAddresses.idleTimeoutInMinutes);
+
           done();
         });
       });
-      it('set should update public ip addresses', function (done) {
-        var cmd = 'network public-ip set -g {group} -n {name} --allocation-method {publicIPAllocationMethodNew} --domain-name-label {domainNameLabelNew} --idle-timeout {idleTimeoutInMinutesNew} --json'.formatArgs(publicIPAddresses);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.equal(0);
-          var output = JSON.parse(result.text);
-          output.name.should.equal(publicIPAddresses.name);
-          output.publicIPAllocationMethod.toLowerCase().should.equal(publicIPAddresses.publicIPAllocationMethodNew.toLowerCase());
-          output.dnsSettings.domainNameLabel.toLowerCase().should.equal(publicIPAddresses.domainNameLabelNew.toLowerCase());
-          output.idleTimeoutInMinutes.should.equal(parseInt(publicIPAddresses.idleTimeoutInMinutesNew, 10));
-          done();
-        });
-      });
+
       it('list should display all public ip addresses in resource group', function (done) {
         var cmd = 'network public-ip list -g {group} --json'.formatArgs(publicIPAddresses);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var outputs = JSON.parse(result.text);
+
           _.some(outputs, function (output) {
             return output.name === publicIPAddresses.name;
           }).should.be.true;
+
           done();
         });
       });
+
       it('delete should delete public ip addresses', function (done) {
         var cmd = 'network public-ip delete -g {group} -n {name} --quiet --json'.formatArgs(publicIPAddresses);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
@@ -204,79 +217,59 @@ describe('arm', function () {
           });
         });
       });
-      it('create with defaults should create public ip addresses with default values', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --location {location} --json'.formatArgs(publicIPAddressesDefault);
+
+      it('create vmss should create virtual machine scale set', function (done) {
+        vmTestUtil.createVMSSWithParamFile(publicIPAddresses, vmTestUtilCreateVMSSWithParamFile, groupName, suite, function (result) {
+          result.exitStatus.should.equal(0);
+          done();
+        });
+      });
+
+      it('show vmss should display public ip address in vmss', function (done) {
+        var cmd = ('network public-ip show -g {group} -n {publicIpAddressName} ' +
+          '--vmss-name {name} --vm-index {virtualmachineIndex} --nic-name {networkInterfaceName} ' +
+          '--ip-config-name {ipConfigurationName} --json').formatArgs(publicIPAddresses);
+         testUtils.executeCommand(suite, retry, cmd, function (result) {
+          result.exitStatus.should.equal(0);
+          var output = JSON.parse(result.text);
+
+          output.name.should.equal(vmTestUtilCreateVMSSWithParamFile.publicIpName);
+          output.dnsSettings.should.not.be.empty;
+          output.dnsSettings.domainNameLabel.should.containEql(vmTestUtilCreateVMSSWithParamFile.domainNameLabel);
+
+          done();
+        });
+      });
+
+      it('list vmss should display all public ip addresses in vmss', function (done) {
+        var cmd = ('network public-ip list -g {group} --vmss-name {name} --json').formatArgs(publicIPAddresses);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
-          output.name.should.equal(publicIPAddressesDefault.name);
-          output.publicIPAllocationMethod.toLowerCase().should.equal(publicIPAddressesDefault.publicIPAllocationMethod.toLowerCase());;
-          output.publicIPAddressVersion.toLowerCase().should.equal(publicIPAddressesDefault.publicIPAddressVersion.toLowerCase());;
-          output.idleTimeoutInMinutes.should.equal(parseInt(publicIPAddressesDefault.idleTimeoutInMinutes, 10));
 
-          cmd = 'network public-ip delete -g {group} -n {name} --quiet --json'.formatArgs(publicIPAddressesDefault);
-          testUtils.executeCommand(suite, retry, cmd, function (result) {
-            result.exitStatus.should.equal(0);
-            done();
-          });
-        });
-      });
-      it('create should fail for ip allocation method out of range', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --allocation-method {publicIPAllocationMethod} --location {location} --json'.formatArgs(ipAllocationMethodOutOfRange);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
+          output.length.should.be.equal(vmssParams.sku.capacity);
+
+          _.some(output, function (publicIpAddress) {
+            return publicIpAddress.name === vmTestUtilCreateVMSSWithParamFile.publicIpName;
+          }).should.be.true;
+
           done();
         });
       });
-      it('create should fail for ip version out of range', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --ip-version {publicIPAddressVersion} --location {location} --json'.formatArgs(ipVersionOutOfRange);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
-          done();
-        });
-      });
-      it('create should fail for invalid symbols in lable', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --domain-name-label {domainNameLabel} --location {location} --json'.formatArgs(invalidSymbolsInLable);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
-          done();
-        });
-      });
-      it('create should fail for idle timeout under range', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --idle-timeout {idleTimeoutInMinutes} --location {location} --json'.formatArgs(idleTimeoutUnderRange);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
-          done();
-        });
-      });
-      it('create should fail for idle timeout over range', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --idle-timeout {idleTimeoutInMinutes} --location {location} --json'.formatArgs(idleTimeoutOverRange);
-        testUtils.executeCommand(suite, retry, cmd, function (result) {
-          result.exitStatus.should.not.equal(0);
-          done();
-        });
-      });
-      it('create should pass for delete of domain name label', function (done) {
-        var cmd = 'network public-ip create -g {group} -n {name} --domain-name-label {domainNameLabel} --location {location} --json'.formatArgs(deleteOfDomainNameLabel);
+
+      it('list vmss index should display all public ip addresses in vmss index', function (done) {
+        var cmd = ('network public-ip list -g {group} --vmss-name {name} ' +
+          '--vm-index {virtualmachineIndex} --nic-name {networkInterfaceName} ' +
+          '--ip-config-name {ipConfigurationName} --json').formatArgs(publicIPAddresses);
         testUtils.executeCommand(suite, retry, cmd, function (result) {
           result.exitStatus.should.equal(0);
           var output = JSON.parse(result.text);
-          output.name.should.equal(deleteOfDomainNameLabel.name);
-          output.dnsSettings.domainNameLabel.toLowerCase().should.equal(deleteOfDomainNameLabel.domainNameLabel.toLowerCase());
 
-          cmd = 'network public-ip set -g {group} -n {name} --domain-name-label {domainNameLabelNew} --json'.formatArgs(deleteOfDomainNameLabel);
-          testUtils.executeCommand(suite, retry, cmd, function (result) {
-            result.exitStatus.should.equal(0);
-            var output = JSON.parse(result.text);
-            output.name.should.equal(deleteOfDomainNameLabel.name);
-            should.not.exist(output.dnsSettings);
+          _.some(output, function (publicIpAddress) {
+            return publicIpAddress.name === vmTestUtilCreateVMSSWithParamFile.publicIpName;
+          }).should.be.true;
 
-            cmd = 'network public-ip delete -g {group} -n {name} --quiet --json'.formatArgs(deleteOfDomainNameLabel);
-            testUtils.executeCommand(suite, retry, cmd, function (result) {
-              result.exitStatus.should.equal(0);
-              done();
-            });
-          });
+          done();
         });
       });
     });
