+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_create_should_create_public_ip_addresses.nock.js
