+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_create_should_fail_for_invalid_symbols_in_lable.nock.js
