+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_create_should_fail_for_ip_allocation_method_out_of_range.nock.js
