+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_create_should_pass_for_delete_of_domain_name_label.nock.js
