+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_create_with_defaults_should_create_public_ip_addresses_with_default_values.nock.js
