+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_delete_should_delete_public_ip_addresses.nock.js
