+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_set_should_update_public_ip_addresses.nock.js
