+++ test/recordings/arm-network-public-ip-tests-generated/arm_network_public_ip_addresses_show_should_display_public_ip_addresses_details.nock.js
