+++ test/recordings/arm-network-public-ip-tests-generated/suite.arm-network-public-ip-tests-generated.nock.js
