+++ test/recordings/arm-network-public-ip-tests/arm_network_public_ip_addresses_custom_create_should_create_public_ip_addresses.nock.js
