+++ test/recordings/arm-network-public-ip-tests/arm_network_public_ip_addresses_custom_create_vmss_should_create_virtual_machine_scale_set.nock.js
