+++ test/recordings/arm-network-public-ip-tests/arm_network_public_ip_addresses_custom_delete_should_delete_public_ip_addresses.nock.js
