+++ test/recordings/arm-network-public-ip-tests/arm_network_public_ip_addresses_custom_list_should_display_all_public_ip_addresses_in_resource_group.nock.js
