+++ test/recordings/arm-network-public-ip-tests/arm_network_public_ip_addresses_custom_list_vmss_index_should_display_all_public_ip_addresses_in_vmss_index.nock.js
