+++ test/recordings/arm-network-public-ip-tests/arm_network_public_ip_addresses_custom_show_vmss_should_display_public_ip_address_in_vmss.nock.js
