+++ test/testlist-arm-network.txt
@@ -68,8 +68,9 @@ commands/arm/network/generated/arm.network.nic-ip-config-tests.text-output.js
 commands/arm/network/generated/arm.network.nic-tests.json-output.js
 commands/arm/network/generated/arm.network.nic-tests.text-output.js
 # Public IP Addresses
-commands/arm/network/arm.network.public-ip-custom-tests.js
 commands/arm/network/arm.network.public-ip-tests.js
+commands/arm/network/generated/arm.network.public-ip-tests.json-output.js
+commands/arm/network/generated/arm.network.public-ip-tests.text-output.js
 # Route Table
 commands/arm/network/arm.network.route-table-tests.js
 commands/arm/network/arm.network.route-table-route-tests.js
